On OpenBSD, search the PATH to find the Bazel executable.
