+++ src/main/cpp/blaze_util_posix.cc
@@ -54,6 +54,7 @@
 #include "src/main/cpp/util/numbers.h"
 #include "src/main/cpp/util/path.h"
 #include "src/main/cpp/util/path_platform.h"
+#include "src/main/cpp/util/strings.h"
 
 namespace blaze {
 
@@ -216,6 +217,29 @@ string GetProcessIdAsString() {
 string GetHomeDir() { return GetPathEnv("HOME"); }
 
 string GetJavaBinaryUnderJavabase() { return "bin/java"; }
+
+string Which(const string& executable) {
+  string path = GetPathEnv("PATH");
+  if (path.empty()) {
+    return "";
+  }
+
+  vector<string> pieces = blaze_util::Split(path, ':');
+  for (string piece : pieces) {
+    if (piece.empty()) {
+      piece = ".";
+    }
+
+    struct stat file_stat;
+    string candidate = blaze_util::JoinPath(piece, executable);
+    if (access(candidate.c_str(), X_OK) == 0 &&
+        stat(candidate.c_str(), &file_stat) == 0 &&
+        (S_ISREG(file_stat.st_mode) || S_ISLNK(file_stat.st_mode))) {
+      return candidate;
+    }
+  }
+  return "";
+}
 
 // Converter of C++ data structures to a C-style array of strings.
 //
